Resolve "two-stage timer necessary for IC time as well"

Closes #33

See merge request lucidac/firmware/hybrid-controller!64

+++ lib/computation/src/run/run.h
@@ -33,7 +33,7 @@ public:
  **/
 class RunConfig {
 public:
-  unsigned int ic_time = 100'000;            ///< Requested Initial Conditions time in Nanoseconds
+  unsigned long long ic_time = 100'000;      ///< Requested Initial Conditions time in Nanoseconds
   unsigned long long op_time = 500'000'000;  ///< Requested Operations time in Nanoseconds
   bool halt_on_overload = true;              ///< Whether to halt the computation on an overload condition
   bool no_streaming = false;                 ///< Request traditional inexact non-streaming run (Non-FlexIO/DMA)
